requirements.txt
Django==2.2.10
djangorestframework==3.10.2
psycopg2==2.8.0
django-notifications-hq==1.6.0
Pillow==7.1.1
django-extra-fields==2.0.2